Changed the server modules - split them into different files in the server/core folder, will put future app modules there
